splitting actions.ts into several files, changing AI retrieval of examples from pool of 5 to 10, troubleshooting perplexity ai prompting to get better quality results
